+++ src/main.rs
@@ -3,5 +3,5 @@ mod simulation_control;
 mod GUI;
 
 pub fn main(){
-    GUI::topology_display::main();
+    GUI::login_window::main();
 }
